// packages/core/src/types.d.ts
export interface DictionaryEntry {
    word: string | RegExp;
    severity: "low" | "medium" | "high";
    category?: "profanity" | "insults" | "spam" | "custom" | string;
}
export interface Dictionary {
    name: string;
    language: string;
    version: string;
    words: DictionaryEntry[];
}
export interface FilterOptions {
    dictionaries?: Dictionary[];
    customWords?: (string | RegExp)[];
    whitelist?: string[];
    mask?: string;
    leetspeakMapping?: Record<string, string[]>;
    faLookalikesMapping?: Record<string, string>;
}
export interface Match {
    word: string;
    matchedText: string;
    start: number;
    end: number;
}
export interface FilterResult {
    originalText: string;
    filteredText: string;
    matches: Match[];
}
export interface TextGuardInstance {
    hasBadWord(text: string): boolean;
    findBadWords(text: string): Match[];
    filter(text: string): FilterResult;
}
//# sourceMappingURL=types.d.ts.map